feat: add news service module
